common iot message context
