+++ mycroft/skills/common_play_skill.py
@@ -143,7 +143,7 @@ class CommonPlaySkill(MycroftSkill, ABC):
         # Stop any currently playing audio
         if self.audioservice.is_playing:
             self.audioservice.stop()
-        self.bus.emit(Message("mycroft.stop"))
+        self.bus.emit(message.forward("mycroft.stop"))
 
         # Save for CPS_play() later, e.g. if phrase includes modifiers like
         # "... on the chromecast"
